requirements.txt
brother_ql
bottle
jinja2
pillow
